+++ nextjs/src/app/api/posts/route.ts
@@ -1,7 +1,7 @@
 export const dynamic = 'force-static';
 
-import { getPublishedPosts, createPost } from "@/datasource"
-import { IPost } from "@/post.entity";
+import { getPublishedPosts, createPost } from "@/infrastructure"
+import { Post } from "@/domain";
 import { NextRequest } from "next/server";
 import { DeepPartial } from "typeorm";
 
@@ -13,7 +13,7 @@ export async function GET() {
 
 export async function POST(request: NextRequest) {
     const requestBody = await request.json();
-    const newPost: DeepPartial<IPost> = {
+    const newPost: DeepPartial<Post> = {
         title: requestBody.title,
         body: requestBody.body,
     };
