+++ nextjs/src/app/posts/new/form.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { createPost, getPostsCount } from "@/datasource";
-import { IPost } from "@/post.entity";
+import { createPost, getPostsCount } from "@/infrastructure";
+import { Post } from "@/domain";
 import { use, useState } from "react";
 import { DeepPartial } from "typeorm";
 
@@ -17,7 +17,7 @@ export default function PostForm({
         const newPost = {
             title: formData.get("title"),
             body: formData.get("body"),
-        } as DeepPartial<IPost>;
+        } as DeepPartial<Post>;
         await createPost(newPost);
         setCnt(await getPostsCount());
     }
